Correction du statut tdb dans le fichier de synthèse
