+++ server/src/common/actions/streamCroisementVoeux.js
@@ -4,7 +4,7 @@ const { findDossier } = require("./findDossier.js");
 const { dateAsString, capitalizeFirstLetter } = require("../utils/stringUtils.js");
 const { sortDescending } = require("../utils/dateUtils.js");
 const { ouiNon } = require("../utils/csvUtils.js");
-const { isNil } = require("lodash");
+const { findRegionByName } = require("../regions.js");
 
 function besoinAide(statut) {
   return !["apprenti", "inscrit"].includes(statut);
@@ -38,7 +38,14 @@ function getDidaskModules(statut) {
   };
 }
 
-function getTrajectoiresPro(statut, voeu) {
+function getJeuneStatut(voeu) {
+  const academies = findRegionByName("Centre-Val de Loire").academies;
+  return academies.find((a) => a.code === voeu.academie.code)
+    ? ouiNon(voeu._meta.jeune_uniquement_en_apprentissage)
+    : "ND";
+}
+
+function getTrajectoiresProUrl(statut, voeu) {
   if (!besoinAide(statut)) {
     return "";
   }
@@ -91,11 +98,9 @@ async function streamCroisementVoeux(options = {}) {
           "Statut dans le tableau de bord": statut ? capitalizeFirstLetter(statut) : "Non trouvé",
           "Date de téléchargement du voeu par l'OF": downloadDate ? dateAsString(downloadDate) : "",
           "La Bonne Alternance": getWidgetLBAUrl(statut, voeu),
-          InserJeunes: getTrajectoiresPro(statut, voeu),
+          InserJeunes: getTrajectoiresProUrl(statut, voeu),
           ...getDidaskModules(statut),
-          "Jeunes uniquement en apprentissage": isNil(voeu._meta.jeune_uniquement_en_apprentissage)
-            ? "ND"
-            : ouiNon(voeu._meta.jeune_uniquement_en_apprentissage),
+          "Jeunes uniquement en apprentissage": getJeuneStatut(voeu),
         };
       },
       { parallel: 10 }
