+++ server/tests/http/csaioRoutes-test.js
@@ -68,7 +68,7 @@ describe("csaioRoutes", () => {
     const { httpClient, createAndLogUser } = await startServer();
     const { auth } = await createAndLogUser("csaio", "password", {
       model: Csaio,
-      region: { code: "11", nom: "Île-de-France" },
+      region: { code: "24", nom: "Centre-Val de Loire" },
     });
 
     const date = DateTime.fromISO("2022-07-23T14:00:00.000Z");
@@ -91,8 +91,8 @@ describe("csaioRoutes", () => {
           telephone_portable: "0612345678",
           adresse: {
             ligne_1: "36 rue des lilas",
-            code_postal: "75019",
-            ville: "Paris",
+            code_postal: "45000",
+            ville: "Orléans",
             pays: "FRANCE",
           },
         },
@@ -116,9 +116,9 @@ describe("csaioRoutes", () => {
           libelle: "1CAP2  CUISINE",
           cle_ministere_educatif: "CLE",
         },
-        academie: { code: "01", nom: "Paris" },
+        academie: { code: "18", nom: "Orléans-Tours" },
         _meta: {
-          adresse: "36 rue des lilas 75019 Paris FRANCE",
+          adresse: "36 rue des lilas 45000 Orléans FRANCE",
           jeune_uniquement_en_apprentissage: true,
         },
       }),
@@ -131,8 +131,8 @@ describe("csaioRoutes", () => {
           telephone_portable: "0712345678",
           adresse: {
             ligne_1: "36 rue des lilas",
-            code_postal: "75019",
-            ville: "Paris",
+            code_postal: "45000",
+            ville: "Orléans",
             pays: "FRANCE",
           },
         },
@@ -156,9 +156,9 @@ describe("csaioRoutes", () => {
           libelle: "2CAP2  CUISINE",
           cle_ministere_educatif: "607555K72235467880206761827130152735855-78100#L60",
         },
-        academie: { code: "01", nom: "Paris" },
+        academie: { code: "18", nom: "Orléans-Tours" },
         _meta: {
-          adresse: "36 rue des lilas 75019 Paris FRANCE",
+          adresse: "36 rue des lilas 45000 Orléans FRANCE",
           jeune_uniquement_en_apprentissage: false,
         },
       }),
@@ -171,8 +171,8 @@ describe("csaioRoutes", () => {
           telephone_portable: "0712345678",
           adresse: {
             ligne_1: "36 rue des lilas",
-            code_postal: "75019",
-            ville: "Paris",
+            code_postal: "45000",
+            ville: "Orléans",
             pays: "FRANCE",
           },
         },
@@ -196,9 +196,9 @@ describe("csaioRoutes", () => {
           libelle: "2CAP2  CUISINE",
           cle_ministere_educatif: "607555K72235467880206761827130152735855-78100#L60",
         },
-        academie: { code: "01", nom: "Paris" },
+        academie: { code: "18", nom: "Orléans-Tours" },
         _meta: {
-          adresse: "36 rue des lilas 75019 Paris FRANCE",
+          adresse: "36 rue des lilas 45000 Orléans FRANCE",
         },
       }),
       insertVoeu({
@@ -229,14 +229,14 @@ describe("csaioRoutes", () => {
     assert.strictEqual(
       response.data,
       `Apprenant INE;Apprenant Nom;Apprenant prénom;Apprenant Téléphone Personnel;Apprenant Téléphone Portable;Apprenant Adresse;Apprenant Adresse Code Postal;Apprenant Adresse Ville;Apprenant Adresse Pays;Etablissement Origine UAI;Etablissement Origine Nom;Etablissement Origine CIO;Etablissement Accueil UAI;Etablissement Accueil Nom;Etablissement Accueil CIO;Formation CFD;Formation MEF;Formation Libellé;Académie;Statut dans le tableau de bord;Date de téléchargement du voeu par l'OF;La Bonne Alternance;InserJeunes;Didask - Prendre contact avec un CFA;Didask - Chercher un employeur;Didask - Préparer un entretien avec un employeur;Didask - S'intégrer dans l'entreprise;Jeunes uniquement en apprentissage
-ABCDEF;Dupont;Robert;0112345678;0612345678;36 rue des lilas 75019 Paris FRANCE;75019;Paris;FRANCE;0751234X;Etablissement Origine;0751234A;0751234J;Etablissement Accueil;0751234Y;40025214;2472521431;1CAP2  CUISINE;Paris;Apprenti;2022-07-23;;;;;;;Oui
-GHIJKL;Dupont;Henri;0212345678;0712345678;36 rue des lilas 75019 Paris FRANCE;75019;Paris;FRANCE;0751234X;Etablissement Origine;0751234A;0751234Z;Etablissement Accueil;0751234Y;50025214;2472521431;2CAP2  CUISINE;Paris;Non trouvé;;https://labonnealternance.pole-emploi.fr/recherche-apprentissage?&display=list&page=fiche&type=training&itemId=607555K72235467880206761827130152735855-78100%23L60;https://trajectoires-pro.apprentissage.beta.gouv.fr/api/inserjeunes/formations/0751234Z-50025214.svg;https://dinum-beta.didask.com/courses/demonstration/60abc18c075edf000065c987;https://dinum-beta.didask.com/courses/demonstration/60d21bf5be76560000ae916e;https://dinum-beta.didask.com/courses/demonstration/60d1adbb877dae00003f0eac;https://dinum-beta.didask.com/courses/demonstration/6283bd5ad9c7ae00003ede91;Non
-MNOPQR;Dupont;Jacques;0212345678;0712345678;36 rue des lilas 75019 Paris FRANCE;75019;Paris;FRANCE;0751234X;Etablissement Origine;0751234A;0751234Z;Etablissement Accueil;0751234Y;50025214;2472521431;2CAP2  CUISINE;Paris;Non trouvé;;https://labonnealternance.pole-emploi.fr/recherche-apprentissage?&display=list&page=fiche&type=training&itemId=607555K72235467880206761827130152735855-78100%23L60;https://trajectoires-pro.apprentissage.beta.gouv.fr/api/inserjeunes/formations/0751234Z-50025214.svg;https://dinum-beta.didask.com/courses/demonstration/60abc18c075edf000065c987;https://dinum-beta.didask.com/courses/demonstration/60d21bf5be76560000ae916e;https://dinum-beta.didask.com/courses/demonstration/60d1adbb877dae00003f0eac;https://dinum-beta.didask.com/courses/demonstration/6283bd5ad9c7ae00003ede91;ND
+ABCDEF;Dupont;Robert;0112345678;0612345678;36 rue des lilas 45000 Orléans FRANCE;45000;Orléans;FRANCE;0751234X;Etablissement Origine;0751234A;0751234J;Etablissement Accueil;0751234Y;40025214;2472521431;1CAP2  CUISINE;Orléans-Tours;Apprenti;2022-07-23;;;;;;;Oui
+GHIJKL;Dupont;Henri;0212345678;0712345678;36 rue des lilas 45000 Orléans FRANCE;45000;Orléans;FRANCE;0751234X;Etablissement Origine;0751234A;0751234Z;Etablissement Accueil;0751234Y;50025214;2472521431;2CAP2  CUISINE;Orléans-Tours;Non trouvé;;https://labonnealternance.pole-emploi.fr/recherche-apprentissage?&display=list&page=fiche&type=training&itemId=607555K72235467880206761827130152735855-78100%23L60;https://trajectoires-pro.apprentissage.beta.gouv.fr/api/inserjeunes/formations/0751234Z-50025214.svg;https://dinum-beta.didask.com/courses/demonstration/60abc18c075edf000065c987;https://dinum-beta.didask.com/courses/demonstration/60d21bf5be76560000ae916e;https://dinum-beta.didask.com/courses/demonstration/60d1adbb877dae00003f0eac;https://dinum-beta.didask.com/courses/demonstration/6283bd5ad9c7ae00003ede91;Non
+MNOPQR;Dupont;Jacques;0212345678;0712345678;36 rue des lilas 45000 Orléans FRANCE;45000;Orléans;FRANCE;0751234X;Etablissement Origine;0751234A;0751234Z;Etablissement Accueil;0751234Y;50025214;2472521431;2CAP2  CUISINE;Orléans-Tours;Non trouvé;;https://labonnealternance.pole-emploi.fr/recherche-apprentissage?&display=list&page=fiche&type=training&itemId=607555K72235467880206761827130152735855-78100%23L60;https://trajectoires-pro.apprentissage.beta.gouv.fr/api/inserjeunes/formations/0751234Z-50025214.svg;https://dinum-beta.didask.com/courses/demonstration/60abc18c075edf000065c987;https://dinum-beta.didask.com/courses/demonstration/60d21bf5be76560000ae916e;https://dinum-beta.didask.com/courses/demonstration/60d1adbb877dae00003f0eac;https://dinum-beta.didask.com/courses/demonstration/6283bd5ad9c7ae00003ede91;Non
 `
     );
   });
 
-  it("Vérifie qu'un csaio peut télécharger le fichier de synthese par région avec les statuts du tdb", async () => {
+  it("Vérifie qu'un csaio peut télécharger le fichier de synthese par région", async () => {
     const { httpClient, createAndLogUser } = await startServer();
     const { auth } = await createAndLogUser("csaio", "password", {
       model: Csaio,
@@ -261,6 +261,12 @@ MNOPQR;Dupont;Jacques;0212345678;0712345678;36 rue des lilas 75019 Paris FRANCE;
           },
         },
         academie: { code: "01", nom: "Paris" },
+        etablissement_accueil: {
+          uai: "0751234Z",
+        },
+        formation: {
+          code_formation_diplome: "50025214",
+        },
         _meta: {
           adresse: "36 rue des lilas 75019 Paris FRANCE",
           jeune_uniquement_en_apprentissage: true,
@@ -281,6 +287,12 @@ MNOPQR;Dupont;Jacques;0212345678;0712345678;36 rue des lilas 75019 Paris FRANCE;
           },
         },
         academie: { code: "01", nom: "Paris" },
+        etablissement_accueil: {
+          uai: "0751234Z",
+        },
+        formation: {
+          code_formation_diplome: "50025214",
+        },
         _meta: {
           adresse: "36 rue des lilas 75019 Paris FRANCE",
           jeune_uniquement_en_apprentissage: false,
@@ -301,6 +313,12 @@ MNOPQR;Dupont;Jacques;0212345678;0712345678;36 rue des lilas 75019 Paris FRANCE;
           },
         },
         academie: { code: "01", nom: "Paris" },
+        etablissement_accueil: {
+          uai: "0751234Z",
+        },
+        formation: {
+          code_formation_diplome: "50025214",
+        },
         _meta: {
           adresse: "36 rue des lilas 75019 Paris FRANCE",
         },
@@ -308,6 +326,8 @@ MNOPQR;Dupont;Jacques;0212345678;0712345678;36 rue des lilas 75019 Paris FRANCE;
       insertDossier({
         ine_apprenant: "ABCDEF",
         statut: "abandon",
+        uai_etablissement: "0751234Z",
+        formation_cfd: "50025214",
         _meta: {
           import_dates: [date.toJSDate()],
         },
@@ -315,6 +335,8 @@ MNOPQR;Dupont;Jacques;0212345678;0712345678;36 rue des lilas 75019 Paris FRANCE;
       insertDossier({
         ine_apprenant: "ABCDEF",
         statut: "apprenti",
+        uai_etablissement: "0751234Z",
+        formation_cfd: "50025214",
         _meta: {
           import_dates: [date.toJSDate()],
         },
